refactor: 이벤트 모듈에서 각 계층별로 패키지 분리
