+++ order/src/main/java/com/klp/order/common/event/application/service/EventStoreService.java
@@ -1,9 +1,11 @@
-package com.klp.order.common.event;
+package com.klp.order.common.event.application.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.klp.order.common.event.DomainEvent;
 import com.klp.order.common.event.domain.EventStore;
 import com.klp.order.common.event.domain.EventType;
+import com.klp.order.common.event.infrastructure.repository.EventStoreRepository;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
